refactor: extract Session's cash and cash-link family into session.cash
